Putting my stuff in
Hey,
I got my tv working and stuff. There is a script that regulates audio and video. The mechanic object is a stand-in for the dad. You get hints by saying help, and you change the video by saying next. I plan to make buttons for changing the video, but I already had the voice recognition so I was just using that to test. IDK what the console errors are. I don't think they are from my stuff, Not sure.
Thanks,